feat(core): add initial implementation of servly host building
Adds a class called `ServlyHost` that provides an abstraction on top of the
default .NET `IHostBuilder` which is intended to make creating an application
with Servly easier.

This includes the internal implementation of `IServlyHost` which is a wrapper
around the default `IHost` from .NET. In the future this can be used for error
handling, reporting or cleanup on exit.

This commit also adds `IServlyBuilder` which will be the main method to add
services and modules related to Servly. A couple of extension methods were
also added to this class to easily configure future Servly Services/Modules.