feat(core): add startup service name banner and platform info
Adds the ability to optionally display a Figgle font banner of the service name
on startup as well as optionally displaying some runtime platform information.

To enable this functionality, the `ServiceName` needs to be set in the
applications `appsettings.json` file under the `Servly` key. Then the two
options `DisplayStartupBanner` and `DisplayPlatformInformation` can optionally
be enabled if desired.

While not a requirement yet, `ServiceName` will be a required configuration
value at some point in the near future.