feat(authentication): add AuthenticationContext to pass authentication details into Dependency Injection
This commit adds the initial implementation of `AuthenticationContext` which is for use in scenarios where authentication details need to be passed into DI in domain driven design type scenarios.

It is implemented in a similar way to the disposable scopes in `ILogger` in the default dotnet logging library using `AsyncLocal<T>` to store the current state and then pop the value on disposal of the scope.