feat(authentication): add AspNetCore middleware to set AuthenticationContext from Authentication
This commit adds a authentication package specifically for AspNetCore that implements `IServiceCollection` and `IApplicationBuilder` extensions for adding `AuthenticationContext` into a AspNetCore project and initializing it via a http middleware.

This middleware provides a configurable interface to set an authentication context. It is configurable so that users can subclass `AuthenticationContext` to add their own "interesting" fields and extract them from the authentication.