feat(idempotency): add AspNetCore middleware for implementing idempotent requests
This commit implements the initial version of the AspNetCore idempotency middleware,
this keeps track of marked incoming requests so that they can be "replayed" instead
of being re-executed and can be used to prevent double executions of POST and PUT requests.

This commit also implements the first version of the Redis persistence provider.